Servo & ultrasonic sensor testing
add a delay() right after myservo.write(pos) to smooth out the movement of the servo. and add delay(15) between the the for-loops. Unfortunately, the servo movement is still a bit jittery. It could be an underpowered servo, a noisy and uneven signal (try a 100uF capacitor to smooth out the signals), or less optimized codes.